Major update
* use AndroidStudio/gradle instead of Eclipse/ant
* upgrade to SDK v22 and build tools v22
* manage network connection from CanvasActivity instead of CanvasView
* show a "No host configured" message instead of opening Settings automatically when no host is configured
* use Android 4.4 immersive full-screen (fixes #59)
* use Material design (using support library)
* add setting for dark canvas (closes #54, #73, closes #76)
