--- app-android/AndroidManifest.xml
@@ -1,38 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="at.bitfire.gfxtablet"
-    android:versionCode="3"
-    android:versionName="1.2" >
-
-    <uses-sdk
-        android:minSdkVersion="14"
-        android:targetSdkVersion="14" />
-
-    <uses-permission android:name="android.permission.INTERNET" />
-    <uses-configuration android:reqTouchScreen="stylus"/>
-    <uses-configuration android:reqTouchScreen="finger"/>
-
-    <application
-        android:allowBackup="true"
-        android:icon="@drawable/ic_launcher"
-        android:label="@string/app_name" >
-        <activity
-            android:name="at.bitfire.gfxtablet.CanvasActivity"
-            android:label="@string/app_name"
-            android:screenOrientation="sensorLandscape" >
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-        </activity>
-        <activity
-            android:name="at.bitfire.gfxtablet.SettingsActivity"
-            android:label="@string/title_activity_settings" >
-            <meta-data
-                android:name="android.support.PARENT_ACTIVITY"
-                android:value="at.bitfire.gfxtablet.CanvasActivity" />
-        </activity>
-    </application>
-
-</manifest>
