--- app-android/IDEAS
@@ -1,5 +1,0 @@
-
-Issues URL: https://github.com/rfc2822/GfxTablet/issues/
-
-Possible improvements:
-  * allow more than one network host in settings [issue #13]
