+++ app-android/app/src/main/java/at/bitfire/gfxtablet/SettingsActivity.java
@@ -1,0 +1,21 @@
+package at.bitfire.gfxtablet;
+
+import android.os.Bundle;
+import android.support.v7.app.ActionBarActivity;
+
+public class SettingsActivity extends ActionBarActivity {
+    public static final String
+		KEY_PREF_HOST = "host_preference",
+		KEY_PREF_STYLUS_ONLY = "stylus_only_preference",
+        KEY_DARK_CANVAS = "dark_canvas_preference";
+    
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+
+        setContentView(R.layout.activity_settings);
+    }
+
+}
