+++ app-android/app/src/main/java/at/bitfire/gfxtablet/SettingsFragment.java
@@ -1,0 +1,12 @@
+package at.bitfire.gfxtablet;
+
+import android.os.Bundle;
+import android.preference.PreferenceFragment;
+
+public class SettingsFragment extends PreferenceFragment {
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        addPreferencesFromResource(R.xml.preferences);
+    }
+}
