+++ app-android/app/src/main/res/drawable-mdpi/ic_launcher.png
