+++ app-android/app/src/main/res/drawable-xhdpi/ic_launcher.png
