+++ app-android/app/src/main/res/layout/activity_settings.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:orientation="vertical" android:layout_width="match_parent"
+    android:layout_height="match_parent">
+
+    <fragment
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:name="at.bitfire.gfxtablet.SettingsFragment"
+        android:id="@+id/fragment"
+        android:layout_gravity="center_horizontal|top" />
+</FrameLayout>
