+++ app-android/app/src/main/res/values/styles.xml
@@ -1,0 +1,10 @@
+<resources>
+
+    <!-- Application theme. -->
+    <style name="AppTheme" parent="Base.Theme.AppCompat.Light.DarkActionBar">
+        <item name="colorPrimary">#4825d0</item>
+        <item name="colorPrimaryDark">#755a8e</item>
+        <item name="colorAccent">#f32eac</item>
+    </style>
+
+</resources>
