--- app-android/res/layout/activity_canvas.xml
@@ -1,17 +1,0 @@
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".CanvasActivity" >
-
-    <LinearLayout
-        android:id="@+id/canvas_layout"
-        android:layout_width="fill_parent"
-        android:layout_height="fill_parent"
-        android:layout_alignParentRight="true"
-        android:layout_alignParentTop="true"
-        android:orientation="vertical" >
-
-    </LinearLayout>
-
-</RelativeLayout>
