--- app-android/res/menu/activity_canvas.xml
@@ -1,13 +1,0 @@
-<menu xmlns:android="http://schemas.android.com/apk/res/android" >
-
-    <item
-        android:orderInCategory="100"
-        android:showAsAction="ifRoom"
-        android:title="@string/menu_settings" android:onClick="showSettings"/>
-
-        <item
-        android:orderInCategory="100"
-        android:showAsAction="never"
-        android:title="@string/menu_about" android:onClick="showAbout"/>
-    
-</menu>
