--- app-android/res/values/strings.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-
-    <string name="app_name">GfxTablet</string>
-    <string name="menu_settings">Settings</string>
-    <string name="preferences_host_title">Network host</string>
-    <string name="stylus_preference">Sense stylus only</string>
-    <string name="fullscreen_preference">Fullscreen</string>
-    <string name="menu_about">About / Help</string>
-    <string name="fullscreen_preference_summary">Title bar will only be hidden when hardware menu button is present</string>
-    <string name="stylus_preference_summary">Ignores touch events from fingers</string>
-
-</resources>
