--- app-android/res/values/strings_activity_settings.xml
@@ -1,14 +1,0 @@
-<resources>
-
-    <string name="title_activity_settings">Settings</string>
-
-    <!-- Strings related to Settings -->
-
-
-    <!-- Example General settings -->
-
-    <!-- Example settings for Data & Sync -->
-
-    <!-- Example settings for Notifications -->
-
-</resources>
