--- app-android/res/xml/preference_headers.xml
@@ -1,9 +1,0 @@
-<preference-headers
-        xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <header android:fragment="at.bitfire.gfxtablet.SettingsActivity$NetworkPrefsFragment"
-            android:title="Networking" />
-    <header android:fragment="at.bitfire.gfxtablet.SettingsActivity$DrawingPrefsFragment"
-            android:title="Drawing" />
-    
-</preference-headers>
